Rename component to `BorderRoundButton`
